refractor: Move actions to their own package
