+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/actions/IAction.kt
@@ -1,4 +1,4 @@
-package com.github.recraftedcivilizations.darkcitizens.tasks.actions
+package com.github.recraftedcivilizations.darkcitizens.actions
 
 import com.github.recraftedcivilizations.darkcitizens.dPlayer.DPlayer
 import org.bukkit.entity.Player
