+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/tasks/Task.kt
@@ -7,7 +7,7 @@ import com.github.recraftedcivilizations.darkcitizens.events.ActionCompleteEvent
 import com.github.recraftedcivilizations.darkcitizens.events.TaskCompleteEvent
 import com.github.recraftedcivilizations.darkcitizens.groups.GroupManager
 import com.github.recraftedcivilizations.darkcitizens.jobs.JobManager
-import com.github.recraftedcivilizations.darkcitizens.tasks.actions.IAction
+import com.github.recraftedcivilizations.darkcitizens.actions.IAction
 import net.milkbowl.vault.economy.Economy
 import org.bukkit.Material
 import org.bukkit.boss.BarColor
